mxc_hdmi: Allow using non CEA video modes and forcing DVI mode
When using a HDMI to DVI connector the EDID are not read and the driver
defaults to 1920x1080.

In order to work with DVI monitors, this commits adds a dvi_mode boolean
parameter which allows to force the controller into DVI mode.

It also adds a only_cea parameter which opens the HDMI driver to
accepting non CEA video modes too.

To use it, add a dvi_mode argument to the hdmi_video node in the DT:

&hdmi_video {
        fsl,phy_reg_vlev = <0x0294>;
        fsl,phy_reg_cksymtx = <0x800d>;
        dvi_mode;
        status = "okay";
};

And pass the only_cea parameter in the kernel command line as:

mxc_hdmi.only_cea=0

Signed-off-by: Alex Gonzalez <alex@lindusembedded.com>